% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\title{Sorteio de delineamentos experimentais}
\usage{
design(k, n, des = c("dic", "dbc"))
}
\arguments{
\item{k}{Número de tratamentos.}

\item{n}{Número de repetições, no DIC, ou de blocos, no DBC.}

\item{des}{Delineamento: \code{"dic"} ou \code{"dbc"}.}
}
\value{
Um \code{data.frame} com uma linha por parcela.
}
\description{
Sorteia os tratamentos às parcelas de um delineamento inteiramente
casualizado (DIC) ou de blocos casualizados (DBC).
}
\examples{
design(k = 3, n = 2, des = "dic")
design(k = 3, n = 2, des = "dbc")
}
