% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meupacote-package.R
\docType{package}
\name{meupacote-package}
\alias{meupacote}
\alias{meupacote-package}
\title{meupacote: Funções de Exemplo para um Primeiro Pacote}
\description{
Reúne funções para operações aritméticas e geométricas simples, para sortear delineamentos experimentais e para o teste t de uma média, além de um conjunto de dados de densidade do solo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fernandomayer.github.io/tdr-teste-ci/}
}

}
\author{
\strong{Maintainer}: Maria Silva \email{maria.silva@example.com}

Authors:
\itemize{
  \item Maria Silva \email{maria.silva@example.com}
}

}
\keyword{internal}
