% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{soma}
\alias{soma}
\title{Soma de dois números}
\usage{
soma(x, y)
}
\value{
Um vetor numérico com a soma de \code{x} e \code{y}.
}
\description{
Recebe dois números e devolve a soma deles. Este primeiro parágrafo é a
descrição: diz o que a função faz e para que ela serve.
}
\details{
Este campo recebe os detalhes técnicos, quando houver, ou
explica melhor como usar algum argumento.
}
\examples{
soma(2, 2)

x <- 3
y <- 4
soma(x = x, y = y)

}
\seealso{
\code{\link[=sum]{sum()}}, que soma todos os elementos de um vetor.
}
\author{
Maria Silva
}
